Refactored BottomPopup in JourneyPage
