+++ Car/__tests__/activity/journey/JourneyPage-test.tsx
@@ -126,132 +126,9 @@ test("renders correctly", async () =>
           </View>
         }
         renderHeader={
-          <View
-            style={
-              Array [
-                Object {
-                  "alignSelf": "center",
-                  "paddingHorizontal": "5%",
-                  "width": "100%",
-                },
-                Object {
-                  "backgroundColor": "white",
-                },
-              ]
-            }
-          >
-            <ForwardRef
-              onPress={[Function]}
-              style={
-                Object {
-                  "flexDirection": "row",
-                  "justifyContent": "space-around",
-                  "paddingLeft": 2,
-                }
-              }
-            >
-              <View
-                style={
-                  Object {
-                    "alignItems": "flex-start",
-                    "justifyContent": "space-around",
-                    "margin": 7,
-                    "marginLeft": 13,
-                  }
-                }
-              >
-                <AvatarLogo
-                  size={38.5}
-                />
-              </View>
-              <View
-                style={
-                  Object {
-                    "flex": 5,
-                    "flexDirection": "column",
-                    "justifyContent": "space-around",
-                  }
-                }
-              >
-                <Text
-                  style={
-                    Array [
-                      Object {
-                        "fontFamily": "Open Sans",
-                        "fontSize": 15,
-                        "fontWeight": "700",
-                      },
-                      Object {
-                        "color": "black",
-                      },
-                    ]
-                  }
-                >
-                   
-                  's ride
-                </Text>
-                <View
-                  style={
-                    Object {
-                      "flexDirection": "row",
-                      "justifyContent": "space-between",
-                    }
-                  }
-                >
-                  <Text
-                    style={
-                      Array [
-                        Object {
-                          "fontFamily": "Open Sans",
-                          "fontSize": 13,
-                          "fontWeight": "100",
-                        },
-                        Object {
-                          "color": "#909095",
-                        },
-                      ]
-                    }
-                  />
-                  <Text
-                    style={
-                      Array [
-                        Object {
-                          "fontFamily": "Open Sans",
-                          "fontSize": 13,
-                          "fontWeight": "700",
-                        },
-                        Object {
-                          "color": "#02A2CF",
-                        },
-                      ]
-                    }
-                  >
-                    01/01/1970
-                  </Text>
-                </View>
-              </View>
-            </ForwardRef>
-            <View
-              style={
-                Object {
-                  "height": 27,
-                }
-              }
-            />
-            <Themed.Divider
-              style={
-                Array [
-                  Object {
-                    "height": 1,
-                    "width": "100%",
-                  },
-                  Object {
-                    "backgroundColor": "#C1C1C5",
-                  },
-                ]
-              }
-            />
-          </View>
+          <DriverBlock
+            journey={null}
+          />
         }
         snapPoints={
           Array [
