+++ Car/src/activity/journey/journey-activity/journey-page/ButtonsBlock/ButtonBlockProps.ts
@@ -1,0 +1,10 @@
+import Journey from "../../../../../../models/journey/Journey";
+
+interface ButtonBlockProps {
+    isDriver: boolean,
+    isPassenger: boolean,
+    isRequested: boolean,
+    journey: Journey
+}
+
+export default ButtonBlockProps;
